requirements.txt
# python 3.5 required
Django==2.1.4
Jinja2==2.10
MarkupSafe==1.1.0
Pillow==5.3.0
PyQt5==5.12
PyQt5-sip==4.19.14
Pygments==2.3.0
Send2Trash==1.5.0
backcall==0.1.0
bleach==3.0.2
certifi==2018.11.29
cmake==3.13.1
colorama==0.4.1
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
dlib==19.16.0
entrypoints==0.2.3
html5lib==1.0.1
imutils==0.5.2
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lxml==4.1.1
matplotlib==3.0.2
mistune==0.8.4
mpmath==1.0.0
mysql-connector-python==8.0.13
nbconvert==5.4.0
nbformat==4.4.0
nose==1.3.7
notebook==5.7.2
numpy==1.15.4
opencv-contrib-python==3.4.4.19
opencv-python==3.4.4.19
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pickleshare==0.7.5
pip==19.0.2
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.6.1
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
pywinpty==0.5.5
pyzmq==17.1.2
qtconsole==4.4.3
scipy==1.1.0
setuptools==40.6.2
simplegeneric==0.8.1
six==1.12.0
sympy==1.3
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
wheel==0.32.3
widgetsnbextension==3.4.2
win-unicode-console==0.5